# backend/src/main/java/com/project/recruitment/dto/RegisterRequest.java
package com.project.recruitment.dto;

import com.project.recruitment.domain.UserRole;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
    @NotBlank(message = "Email is required")
    @Email(message = "Invalid email format")
    String email,

    @NotBlank(message = "Password is required")
    @Size(min = 6, message = "Password must be at least 6 characters")
    String password,

    @NotBlank(message = "Full name is required")
    String fullName,

    String phoneNumber,

    @NotNull(message = "Role is required")
    UserRole role
) {
}